move eventButtons to events/dashboard folder
