+++ design-at-cornell/src/events/dashboard/eventButtons.ts
@@ -1,8 +1,8 @@
-import googleCalIcon from '../static/images/google-cal-icon.svg';
-import emailIcon from '../static/images/email-icon.svg';
-import facebookIcon from '../static/images/facebook-icon.svg';
-import zoomIcon from '../static/images/zoom-icon.svg';
-import linkIcon from '../static/images/link-icon.svg';
+import googleCalIcon from '../../static/images/google-cal-icon.svg';
+import emailIcon from '../../static/images/email-icon.svg';
+import facebookIcon from '../../static/images/facebook-icon.svg';
+import zoomIcon from '../../static/images/zoom-icon.svg';
+import linkIcon from '../../static/images/link-icon.svg';
 
 type EventButton = {
   icon: string;
